refactor(quadlet): move generator tests into a tests/ submodule
quadlet/mod.rs had grown to 636 lines, over the 500-line hard limit, almost
entirely from its inline test module. Move the tests next to the code in a
tests/ submodule split by theme — units.rs (network/volume/unit emission,
warnings, injection safety) and fields.rs (container key mapping) — with the
shared lookup helper in tests/mod.rs. mod.rs keeps only the generator. Pure
relocation, no behavior change.
